@@ -2,14 +2,7 @@
   <Import Project="..\..\versions.props" />
 
   <PropertyGroup>
-    <Description>Unit test project for Steeltoe.Management.CloudFoundryCore</Description>
-    <TargetFrameworks>netcoreapp2.0;net461</TargetFrameworks>
-    <RuntimeFrameworkVersion Condition=" '$(TargetFramework)' == 'netcoreapp2.0' ">2.0.3</RuntimeFrameworkVersion>
-    <AssemblyName>Steeltoe.Management.CloudFoundryCore.Test</AssemblyName>
-    <PackageId>Steeltoe.Management.CloudFoundryCore.Test</PackageId>
-    <PackageTags>Spring Cloud;Actuator;Management;Monitoring;CloudFoundry</PackageTags>
-    <PackageProjectUrl>https://steeltoe.io</PackageProjectUrl>
-    <PackageLicenseUrl>https://www.apache.org/licenses/LICENSE-2.0</PackageLicenseUrl>
+    <TargetFrameworks>netcoreapp2.1;net461</TargetFrameworks>
   </PropertyGroup>
   
   <ItemGroup>
