@@ -57,7 +57,14 @@ namespace Steeltoe.Management.Endpoint.CloudFoundry
                 }
                 else
                 {
-                    links._links.Add(opt.Id, new Link(baseUrl + "/" + opt.Id));
+                    if (!string.IsNullOrEmpty(opt.Id) && !links._links.ContainsKey(opt.Id))
+                    {
+                        links._links.Add(opt.Id, new Link(baseUrl + "/" + opt.Id));
+                    }
+                    else if (links._links.ContainsKey(opt.Id))
+                    {
+                        _logger?.LogWarning("Duplicate endpoint id detected: {DuplicateEndpointId}", opt.Id);
+                    }
                 }
             }
 
