@@ -15,64 +15,50 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Primitives;
-using Newtonsoft.Json;
-using Steeltoe.Common.Http;
-using Steeltoe.Management.Endpoint.Security;
+using Steeltoe.Common;
 using System;
-using System.Collections.Generic;
 using System.Net;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Net.Security;
 using System.Threading.Tasks;
 
 namespace Steeltoe.Management.Endpoint.CloudFoundry
 {
     public class CloudFoundrySecurityMiddleware
     {
-        private const int DEFAULT_GETPERMISSIONS_TIMEOUT = 5000;   // Milliseconds
-        private const string APPLICATION_ID_MISSING_MESSAGE = "Application id is not available";
-        private const string ENDPOINT_NOT_CONFIGURED_MESSAGE = "Endpoint is not available";
-        private const string AUTHORIZATION_HEADER_INVALID = "Authorization header is missing or invalid";
-        private const string CLOUDFOUNDRY_API_MISSING_MESSAGE = "Cloud controller URL is not available";
-        private const string CLOUDFOUNDRY_NOT_REACHABLE_MESSAGE = "Cloud controller not reachable";
-        private const string ACCESS_DENIED_MESSAGE = "Access denied";
-        private const string AUTHORIZATION_HEADER = "Authorization";
-        private const string BEARER = "bearer";
-        private const string READ_SENSITIVE_DATA = "read_sensitive_data";
         private RequestDelegate _next;
         private ILogger<CloudFoundrySecurityMiddleware> _logger;
         private ICloudFoundryOptions _options;
+        private SecurityBase _base;
 
         public CloudFoundrySecurityMiddleware(RequestDelegate next, ICloudFoundryOptions options, ILogger<CloudFoundrySecurityMiddleware> logger)
         {
             _next = next;
             _logger = logger;
             _options = options;
+            _base = new SecurityBase(options, logger);
         }
 
         public async Task Invoke(HttpContext context)
         {
             _logger.LogDebug("Invoke({0})", context.Request.Path.Value);
 
-            if (IsCloudFoundryRequest(context))
+            if (Platform.IsCloudFoundry && _options.IsEnabled && _base.IsCloudFoundryRequest(context.Request.Path))
             {
                 if (string.IsNullOrEmpty(_options.ApplicationId))
                 {
-                    await ReturnError(context, new SecurityResult(HttpStatusCode.ServiceUnavailable, APPLICATION_ID_MISSING_MESSAGE));
+                    await ReturnError(context, new SecurityResult(HttpStatusCode.ServiceUnavailable, _base.APPLICATION_ID_MISSING_MESSAGE));
                     return;
                 }
 
                 if (string.IsNullOrEmpty(_options.CloudFoundryApi))
                 {
-                    await ReturnError(context, new SecurityResult(HttpStatusCode.ServiceUnavailable, CLOUDFOUNDRY_API_MISSING_MESSAGE));
+                    await ReturnError(context, new SecurityResult(HttpStatusCode.ServiceUnavailable, _base.CLOUDFOUNDRY_API_MISSING_MESSAGE));
                     return;
                 }
 
                 IEndpointOptions target = FindTargetEndpoint(context.Request.Path);
                 if (target == null)
                 {
-                    await ReturnError(context, new SecurityResult(HttpStatusCode.ServiceUnavailable, ENDPOINT_NOT_CONFIGURED_MESSAGE));
+                    await ReturnError(context, new SecurityResult(HttpStatusCode.ServiceUnavailable, _base.ENDPOINT_NOT_CONFIGURED_MESSAGE));
                     return;
                 }
 
@@ -86,7 +72,7 @@ namespace Steeltoe.Management.Endpoint.CloudFoundry
                 var permissions = sr.Permissions;
                 if (!target.IsAccessAllowed(permissions))
                 {
-                    await ReturnError(context, new SecurityResult(HttpStatusCode.Forbidden, ACCESS_DENIED_MESSAGE));
+                    await ReturnError(context, new SecurityResult(HttpStatusCode.Forbidden, _base.ACCESS_DENIED_MESSAGE));
                     return;
                 }
             }
@@ -96,12 +82,12 @@ namespace Steeltoe.Management.Endpoint.CloudFoundry
 
         internal string GetAccessToken(HttpRequest request)
         {
-            if (request.Headers.TryGetValue(AUTHORIZATION_HEADER, out StringValues headerVal))
+            if (request.Headers.TryGetValue(_base.AUTHORIZATION_HEADER, out StringValues headerVal))
             {
                 string header = headerVal.ToString();
-                if (header.StartsWith(BEARER, StringComparison.OrdinalIgnoreCase))
+                if (header.StartsWith(_base.BEARER, StringComparison.OrdinalIgnoreCase))
                 {
-                    return header.Substring(BEARER.Length + 1);
+                    return header.Substring(_base.BEARER.Length + 1);
                 }
             }
 
@@ -111,69 +97,7 @@ namespace Steeltoe.Management.Endpoint.CloudFoundry
         internal async Task<SecurityResult> GetPermissions(HttpContext context)
         {
             string token = GetAccessToken(context.Request);
-            if (string.IsNullOrEmpty(token))
-            {
-                return new SecurityResult(HttpStatusCode.Unauthorized, AUTHORIZATION_HEADER_INVALID);
-            }
-
-            string checkPermissionsUri = _options.CloudFoundryApi + "/v2/apps/" + _options.ApplicationId + "/permissions";
-            var request = new HttpRequestMessage(HttpMethod.Get, checkPermissionsUri);
-            AuthenticationHeaderValue auth = new AuthenticationHeaderValue("bearer", token);
-            request.Headers.Authorization = auth;
-
-            // If certificate validation is disabled, inject a callback to handle properly
-            HttpClientHelper.ConfigureCertificateValidatation(
-                _options.ValidateCertificates,
-                out SecurityProtocolType prevProtocols,
-                out RemoteCertificateValidationCallback prevValidator);
-            try
-            {
-                _logger.LogDebug("GetPermissions({0}, {1})", checkPermissionsUri, token);
-
-                using (var client = HttpClientHelper.GetHttpClient(_options.ValidateCertificates, DEFAULT_GETPERMISSIONS_TIMEOUT))
-                {
-                    using (HttpResponseMessage response = await client.SendAsync(request))
-                    {
-                        if (response.StatusCode != HttpStatusCode.OK)
-                        {
-                            _logger?.LogInformation(
-                                "Cloud Foundry returned status: {0} while obtaining permissions from: {1}",
-                                response.StatusCode,
-                                checkPermissionsUri);
-
-                            if (response.StatusCode == HttpStatusCode.Forbidden)
-                            {
-                                return new SecurityResult(HttpStatusCode.Forbidden, ACCESS_DENIED_MESSAGE);
-                            }
-                            else
-                            {
-                                return new SecurityResult(HttpStatusCode.ServiceUnavailable, CLOUDFOUNDRY_NOT_REACHABLE_MESSAGE);
-                            }
-                        }
-
-                        return new SecurityResult(await GetPermissions(response));
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                _logger?.LogError(
-                    "Cloud Foundry returned execption: {0} while obtaining permissions from: {1}",
-                        e,
-                        checkPermissionsUri);
-                return new SecurityResult(HttpStatusCode.ServiceUnavailable, CLOUDFOUNDRY_NOT_REACHABLE_MESSAGE);
-            }
-            finally
-            {
-                HttpClientHelper.RestoreCertificateValidation(_options.ValidateCertificates, prevProtocols, prevValidator);
-            }
-        }
-
-        protected internal bool IsCloudFoundryRequest(HttpContext context)
-        {
-            PathString path = new PathString(_options.Path == "/" ? string.Empty : _options.Path);
-            bool startsWith = context.Request.Path.StartsWithSegments(path);
-            return startsWith;
+            return await _base.GetPermissionsAsync(token);
         }
 
         private IEndpointOptions FindTargetEndpoint(PathString path)
@@ -196,56 +120,7 @@ namespace Steeltoe.Management.Endpoint.CloudFoundry
             LogError(context, error);
             context.Response.Headers.Add("Content-Type", "application/json;charset=UTF-8");
             context.Response.StatusCode = (int)error.Code;
-            await context.Response.WriteAsync(Serialize(error));
-        }
-
-        private string Serialize(SecurityResult error)
-        {
-            try
-            {
-                return JsonConvert.SerializeObject(error);
-            }
-            catch (Exception e)
-            {
-                _logger.LogError("Serialization Exception: {0}", e);
-            }
-
-            return string.Empty;
-        }
-
-        private async Task<Permissions> GetPermissions(HttpResponseMessage response)
-        {
-            string json = string.Empty;
-            Permissions permissions = Permissions.NONE;
-
-            try
-            {
-                json = await response.Content.ReadAsStringAsync();
-
-                _logger.LogDebug("GetPermisions returned json: {0}", json);
-
-                var result = JsonConvert.DeserializeObject<Dictionary<string, object>>(json);
-
-                if (result.TryGetValue(READ_SENSITIVE_DATA, out object perm))
-                {
-                    bool boolResult = (bool)perm;
-                    if (boolResult)
-                    {
-                        permissions = Permissions.FULL;
-                    }
-                    else
-                    {
-                        permissions = Permissions.RESTRICTED;
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError("Exception {0} extracting permissions from {1}", e, json);
-            }
-
-            _logger.LogDebug("GetPermisions returning: {0}", permissions);
-            return permissions;
+            await context.Response.WriteAsync(_base.Serialize(error));
         }
 
         private void LogError(HttpContext context, SecurityResult error)
