@@ -14,26 +14,27 @@
 
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
+using Steeltoe.Common.HealthChecks;
 using Steeltoe.Management.Endpoint.Middleware;
 using System.Threading.Tasks;
 
 namespace Steeltoe.Management.Endpoint.Health
 {
-    public class HealthEndpointMiddleware : EndpointMiddleware<Health>
+    public class HealthEndpointMiddleware : EndpointMiddleware<HealthCheckResult>
     {
         private RequestDelegate _next;
 
         public HealthEndpointMiddleware(RequestDelegate next, ILogger<HealthEndpointMiddleware> logger = null)
-            : base(logger)
+            : base(logger: logger)
         {
             _next = next;
         }
 
         public async Task Invoke(HttpContext context, HealthEndpoint endpoint)
         {
-            this.endpoint = endpoint;
+            _endpoint = endpoint;
 
-            if (IsHealthRequest(context))
+            if (RequestVerbAndPathMatch(context.Request.Method, context.Request.Path.Value))
             {
                 await HandleHealthRequestAsync(context);
             }
@@ -45,21 +46,17 @@ namespace Steeltoe.Management.Endpoint.Health
 
         protected internal async Task HandleHealthRequestAsync(HttpContext context)
         {
-            var serialInfo = HandleRequest();
-            logger?.LogDebug("Returning: {0}", serialInfo);
+            var serialInfo = DoRequest(context);
+            _logger?.LogDebug("Returning: {0}", serialInfo);
             context.Response.Headers.Add("Content-Type", "application/vnd.spring-boot.actuator.v1+json");
             await context.Response.WriteAsync(serialInfo);
         }
 
-        protected internal bool IsHealthRequest(HttpContext context)
+        protected internal string DoRequest(HttpContext context)
         {
-            if (!context.Request.Method.Equals("GET"))
-            {
-                return false;
-            }
-
-            PathString path = new PathString(endpoint.Path);
-            return context.Request.Path.Equals(path);
+            var result = _endpoint.Invoke();
+            context.Response.StatusCode = ((HealthEndpoint)_endpoint).GetStatusCode(result);
+            return Serialize(result);
         }
     }
 }
