@@ -15,9 +15,8 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Steeltoe.Management.Endpoint.Middleware;
-using System;
+using Steeltoe.Management.EndpointBase;
 using System.IO;
-using System.IO.Compression;
 using System.Threading.Tasks;
 
 namespace Steeltoe.Management.Endpoint.HeapDump
@@ -27,14 +26,14 @@ namespace Steeltoe.Management.Endpoint.HeapDump
         private RequestDelegate _next;
 
         public HeapDumpEndpointMiddleware(RequestDelegate next, HeapDumpEndpoint endpoint, ILogger<HeapDumpEndpointMiddleware> logger = null)
-            : base(endpoint, logger)
+            : base(endpoint, logger: logger)
         {
             _next = next;
         }
 
         public async Task Invoke(HttpContext context)
         {
-            if (IsHeapDumpRequest(context))
+            if (RequestVerbAndPathMatch(context.Request.Method, context.Request.Path.Value))
             {
                 await HandleHeapDumpRequestAsync(context);
             }
@@ -46,8 +45,8 @@ namespace Steeltoe.Management.Endpoint.HeapDump
 
         protected internal async Task HandleHeapDumpRequestAsync(HttpContext context)
         {
-            var filename = endpoint.Invoke();
-            logger?.LogDebug("Returning: {0}", filename);
+            var filename = _endpoint.Invoke();
+            _logger?.LogDebug("Returning: {0}", filename);
             context.Response.Headers.Add("Content-Type", "application/octet-stream");
 
             if (!File.Exists(filename))
@@ -57,7 +56,7 @@ namespace Steeltoe.Management.Endpoint.HeapDump
             }
 
             string gzFilename = filename + ".gz";
-            var result = await CompresssAsync(filename, gzFilename);
+            var result = await Utils.CompressFileAsync(filename, gzFilename);
 
             if (result != null)
             {
@@ -75,46 +74,6 @@ namespace Steeltoe.Management.Endpoint.HeapDump
             {
                 context.Response.StatusCode = StatusCodes.Status404NotFound;
             }
-        }
-
-        protected internal async Task<Stream> CompresssAsync(string filename, string gzFilename)
-        {
-            try
-            {
-                using (var input = new FileStream(filename, FileMode.Open))
-                {
-                    using (var output = new FileStream(gzFilename, FileMode.CreateNew))
-                    {
-                        using (var gzipStream = new GZipStream(output, CompressionLevel.Fastest))
-                        {
-                            await input.CopyToAsync(gzipStream);
-                        }
-                    }
-                }
-
-                return new FileStream(gzFilename, FileMode.Open);
-            }
-            catch (Exception e)
-            {
-                logger?.LogError(e, "Unable to compress dump");
-            }
-            finally
-            {
-                File.Delete(filename);
-            }
-
-            return null;
-        }
-
-        protected internal bool IsHeapDumpRequest(HttpContext context)
-        {
-            if (!context.Request.Method.Equals("GET"))
-            {
-                return false;
-            }
-
-            PathString path = new PathString(endpoint.Path);
-            return context.Request.Path.Equals(path);
         }
     }
 }
