@@ -20,19 +20,19 @@ using System.Threading.Tasks;
 
 namespace Steeltoe.Management.Endpoint.Trace
 {
-    public class TraceEndpointMiddleware : EndpointMiddleware<List<Trace>>
+    public class TraceEndpointMiddleware : EndpointMiddleware<List<TraceResult>>
     {
         private RequestDelegate _next;
 
         public TraceEndpointMiddleware(RequestDelegate next, TraceEndpoint endpoint, ILogger<TraceEndpointMiddleware> logger = null)
-            : base(endpoint, logger)
+            : base(endpoint, logger: logger)
         {
             _next = next;
         }
 
         public async Task Invoke(HttpContext context)
         {
-            if (IsTraceRequest(context))
+            if (RequestVerbAndPathMatch(context.Request.Method, context.Request.Path.Value))
             {
                 await HandleTraceRequestAsync(context);
             }
@@ -45,20 +45,9 @@ namespace Steeltoe.Management.Endpoint.Trace
         protected internal async Task HandleTraceRequestAsync(HttpContext context)
         {
             var serialInfo = HandleRequest();
-            logger?.LogDebug("Returning: {0}", serialInfo);
+            _logger?.LogDebug("Returning: {0}", serialInfo);
             context.Response.Headers.Add("Content-Type", "application/vnd.spring-boot.actuator.v1+json");
             await context.Response.WriteAsync(serialInfo);
-        }
-
-        protected internal bool IsTraceRequest(HttpContext context)
-        {
-            if (!context.Request.Method.Equals("GET"))
-            {
-                return false;
-            }
-
-            PathString path = new PathString(endpoint.Path);
-            return context.Request.Path.Equals(path);
         }
     }
 }
