@@ -48,24 +48,6 @@ namespace Steeltoe.Management.Endpoint.Info.Test
         };
 
         [Fact]
-        public void IsInfoRequest_ReturnsExpected()
-        {
-            var opts = new InfoOptions();
-            var contribs = new List<IInfoContributor>() { new GitInfoContributor() };
-            var ep = new InfoEndpoint(opts, contribs);
-            var middle = new InfoEndpointMiddleware(null, ep);
-
-            var context = CreateRequest("GET", "/info");
-            Assert.True(middle.IsInfoRequest(context));
-
-            var context2 = CreateRequest("PUT", "/info");
-            Assert.False(middle.IsInfoRequest(context2));
-
-            var context3 = CreateRequest("GET", "/badpath");
-            Assert.False(middle.IsInfoRequest(context3));
-        }
-
-        [Fact]
         public async void HandleInfoRequestAsync_ReturnsExpected()
         {
             var opts = new InfoOptions();
@@ -122,6 +104,19 @@ namespace Steeltoe.Management.Endpoint.Info.Test
                 Assert.True(gitNode.ContainsKey("remote"));
                 Assert.True(gitNode.ContainsKey("tags"));
             }
+        }
+
+        [Fact]
+        public void InfoEndpointMiddleware_PathAndVerbMatching_ReturnsExpected()
+        {
+            var opts = new InfoOptions();
+            var contribs = new List<IInfoContributor>() { new GitInfoContributor() };
+            var ep = new InfoEndpoint(opts, contribs);
+            var middle = new InfoEndpointMiddleware(null, ep);
+
+            Assert.True(middle.RequestVerbAndPathMatch("GET", "/info"));
+            Assert.False(middle.RequestVerbAndPathMatch("PUT", "/info"));
+            Assert.False(middle.RequestVerbAndPathMatch("GET", "/badpath"));
         }
 
         private HttpContext CreateRequest(string method, string path)
